implementando id automaticos e validacoes em endpoints

+++ src/business/PostBusiness.ts
@@ -1,13 +1,15 @@
 import { PostDatabase } from "../database/PostDatabase"
-import { PostCreateInputDTO} from "../dtos/postCreate.dto"
+import { PostCreateInputDTO } from "../dtos/postCreate.dto"
+import { PostDeleteInputDTO } from "../dtos/postDelete.dto"
 import { PostGetAllOutputDTO } from "../dtos/postGetAll.dto"
 import { PostUpdateInputDTO } from "../dtos/postUpdate.dto"
+import { TokenCheckInputDTO } from "../dtos/tokenCheck.dto"
 import { BadRequestError } from "../errors/BadRequestError"
-import { Post } from "../models/Post"
+import { USER_ROLES } from "../models/User"
 import { IdGenerator } from "../service/IdGenerator"
 import { TokenManager } from "../service/TokenManager"
 import { PostDB } from "../types"
-import { uuid } from "uuidv4"
+
 
 export class PostBusiness {
     constructor(
@@ -17,14 +19,14 @@ export class PostBusiness {
     ) { }
 
 
-    public create = async (input: PostCreateInputDTO): Promise<void>  => {
+    public create = async (input: PostCreateInputDTO): Promise<void> => {
         const id = this.idGenerator.generate()
 
-        const {content, token} = input
+        const { content, token } = input
 
         const result = this.tokenManager.getPayload(token)
 
-        if(!result){
+        if (!result) {
             throw new Error("token invalido.")
         }
 
@@ -41,7 +43,15 @@ export class PostBusiness {
         await this.postDatabase.createPost(postCreateDB)
     }
 
-    public getAll = async (): Promise<PostGetAllOutputDTO[]> => {
+    public getAll = async (input: TokenCheckInputDTO): Promise<PostGetAllOutputDTO[]> => {
+
+        const { token } = input
+
+        const payload = this.tokenManager.getPayload(token)
+
+        if (payload === null) {
+            throw new BadRequestError("token invalido.")
+        }
 
         const result = await this.postDatabase.findPost()
 
@@ -65,21 +75,46 @@ export class PostBusiness {
 
         const {
             id,
-            content
+            content,
+            token
         } = input
 
+        const payload = this.tokenManager.getPayload(token)
+        if (payload === null) {
+            throw new BadRequestError("token invalido.")
+        }
+        const postDB = await this.postDatabase.findPostById(id)
+
+        if (!postDB) {
+            throw new BadRequestError("id invalido.")
+        }
+
+
+        if (payload.id !== postDB.creator_id) {
+            throw new BadRequestError("id incorreto.")
+        }
 
         await this.postDatabase.updatePost(id, content)
     }
 
-    public delete = async (id: string) => {
+    public delete = async (input: PostDeleteInputDTO) => {
 
+        const { id, token } = input
+
+        const payload = this.tokenManager.getPayload(token)
+        if (payload === null) {
+            throw new BadRequestError("token invalido.")
+        }
 
         const result = await this.postDatabase.findPostById(id)
 
         if (!result) {
             throw new BadRequestError("'Id' não encontrado.")
         }
-        await this.postDatabase.delete(id)
+
+        if (payload.id === result.id || payload.role === USER_ROLES.ADMIN) {
+            await this.postDatabase.delete(id)
+        }
+
     }
 }
